remove transactions
