+++ execution_engine/src/storage.rs
@@ -6,8 +6,6 @@ pub mod error;
 pub mod global_state;
 /// Store module.
 pub mod store;
-/// Transaction Source.
-pub mod transaction_source;
 /// Merkle Trie implementation.
 pub mod trie;
 /// Merkle Trie storage.
@@ -15,10 +13,4 @@ pub mod trie_store;
 
 const MAX_DBS: u32 = 2;
 
-pub use transaction_source::ROCKS_DB_DATA_DIR;
-
-#[cfg(test)]
-pub(crate) const DEFAULT_TEST_MAX_DB_SIZE: usize = 52_428_800; // 50 MiB
-
-#[cfg(test)]
-pub(crate) const DEFAULT_TEST_MAX_READERS: u32 = 512;
+pub use trie_store::db::ROCKS_DB_DATA_DIR;
