+++ execution_engine/src/storage/trie_store/db.rs
@@ -2,61 +2,48 @@
 
 use std::{
     collections::HashMap,
+    path::{Path, PathBuf},
     sync::{Arc, Mutex},
 };
 
-use casper_types::{bytesrepr, Key, StoredValue};
-use lmdb::{Database, DatabaseFlags};
+use rocksdb::{
+    BlockBasedOptions, BoundColumnFamily, DBIteratorWithThreadMode, DBWithThreadMode, IteratorMode,
+    MultiThreaded, Options,
+};
+
+use casper_types::{bytesrepr, bytesrepr::Bytes, Key, StoredValue};
 
 use casper_hashing::Digest;
 
 use crate::storage::{
     error,
     global_state::CommitError,
-    store::Store,
-    transaction_source::db::{LmdbEnvironment, RocksDbStore},
+    store::{ErrorSource, Readable, Store, Writable},
     trie::Trie,
-    trie_store::{self, TrieStore},
+    trie_store::TrieStore,
 };
 
-/// An LMDB-backed trie store.
-///
-/// Wraps [`lmdb::Database`].
-#[derive(Debug, Clone)]
-pub struct LmdbTrieStore {
-    db: Database,
-}
+/// Relative location (to storage) where rocksdb data will be stored.
+pub const ROCKS_DB_DATA_DIR: &str = "rocksdb-data";
 
-impl LmdbTrieStore {
-    /// Constructor for new `LmdbTrieStore`.
-    pub fn new(
-        env: &LmdbEnvironment,
-        maybe_name: Option<&str>,
-        flags: DatabaseFlags,
-    ) -> Result<Self, error::Error> {
-        let name = Self::name(maybe_name);
-        let db = env.env().create_db(Some(&name), flags)?;
-        Ok(LmdbTrieStore { db })
-    }
+const ROCKS_DB_BLOCK_SIZE_BYTES: usize = 256 * 1024;
+const ROCKS_DB_COMPRESSION_TYPE: rocksdb::DBCompressionType = rocksdb::DBCompressionType::Zstd;
+const ROCKS_DB_COMPACTION_STYLE: rocksdb::DBCompactionStyle = rocksdb::DBCompactionStyle::Level;
+const ROCKS_DB_ZSTD_MAX_DICT_BYTES: i32 = 256 * 1024;
+const ROCKS_DB_MAX_LEVEL_FILE_SIZE_BYTES: u64 = 512 * 1024 * 1024;
+const ROCKS_DB_MAX_OPEN_FILES: i32 = 768;
+const ROCKS_DB_ZSTD_COMPRESSION_LEVEL: i32 = 3; // Default compression level
+const ROCKS_DB_ZSTD_STRATEGY: i32 = 4; // 4: Lazy
+const ROCKS_DB_WINDOW_BITS: i32 = -14;
 
-    /// Constructor for `LmdbTrieStore` which opens an existing lmdb store file.
-    pub fn open(env: &LmdbEnvironment, maybe_name: Option<&str>) -> Result<Self, error::Error> {
-        let name = Self::name(maybe_name);
-        let db = env.env().open_db(Some(&name))?;
-        Ok(LmdbTrieStore { db })
-    }
+/// Column family name for the v1 trie data store.
+const TRIE_V1_COLUMN_FAMILY: &str = "trie_v1_column";
 
-    fn name(maybe_name: Option<&str>) -> String {
-        maybe_name
-            .map(|name| format!("{}-{}", trie_store::NAME, name))
-            .unwrap_or_else(|| String::from(trie_store::NAME))
-    }
+/// Column family name for tracking progress of data migration.
+const LMDB_MIGRATED_STATE_ROOTS_COLUMN_FAMILY: &str = "lmdb_migrated_state_roots_column";
 
-    /// Get a handle to the underlying database.
-    pub fn get_db(&self) -> Database {
-        self.db
-    }
-}
+/// Working set column famliy.
+const WORKING_SET_COLUMN_FAMILY: &str = "working_set";
 
 /// Cache used by the scratch trie.  The keys represent the hash of the trie being cached.  The
 /// values represent:  1) A boolean, where `false` means the trie was _not_ written and `true` means
@@ -155,6 +142,215 @@ impl Store<Digest, Trie<Key, StoredValue>> for ScratchTrieStore {
                 }
             }
         }
+    }
+}
+
+/// Represents the state of a migration of a state root from lmdb to rocksdb.
+pub enum RootMigration {
+    /// Has the migration been left incomplete
+    NotMigrated,
+    /// Has the migration been not yet been started or completed
+    Incomplete,
+    /// Has the migration been completed
+    Complete,
+}
+
+impl RootMigration {
+    /// Has the migration been left incomplete
+    pub fn is_incomplete(&self) -> bool {
+        matches!(self, RootMigration::Incomplete)
+    }
+    /// Has the migration been not yet been started or completed
+    pub fn is_not_migrated(&self) -> bool {
+        matches!(self, RootMigration::NotMigrated)
+    }
+    /// Has the migration been completed
+    pub fn is_complete(&self) -> bool {
+        matches!(self, RootMigration::Complete)
+    }
+}
+
+impl ErrorSource for ScratchTrieStore {
+    type Error = error::Error;
+}
+
+impl Readable for ScratchTrieStore {
+    fn read(&self, key: &[u8]) -> Result<Option<Bytes>, Self::Error> {
+        self.store.read(key)
+    }
+}
+
+impl Writable for ScratchTrieStore {
+    fn write(&self, key: &[u8], value: &[u8]) -> Result<(), Self::Error> {
+        self.store.write(key, value)
+    }
+}
+
+impl ErrorSource for RocksDbStore {
+    type Error = error::Error;
+}
+
+impl Readable for RocksDbStore {
+    fn read(&self, key: &[u8]) -> Result<Option<Bytes>, Self::Error> {
+        let cf = self.trie_column_family()?;
+        Ok(self.db.get_cf(&cf, key)?.map(|some| {
+            let value = some.as_ref();
+            Bytes::from(value)
+        }))
+    }
+}
+
+impl Writable for RocksDbStore {
+    fn write(&self, key: &[u8], value: &[u8]) -> Result<(), Self::Error> {
+        let cf = self.trie_column_family()?;
+        let _result = self.db.put_cf(&cf, key, value)?;
+        Ok(())
+    }
+}
+
+/// Environment for rocksdb.
+#[derive(Clone)]
+pub struct RocksDbStore {
+    db: Arc<DBWithThreadMode<MultiThreaded>>,
+    pub(crate) path: PathBuf,
+}
+
+const ACTIVE_COLUMN_FAMILIES: &[&str] = &[
+    TRIE_V1_COLUMN_FAMILY,
+    LMDB_MIGRATED_STATE_ROOTS_COLUMN_FAMILY,
+    WORKING_SET_COLUMN_FAMILY,
+];
+
+impl RocksDbStore {
+    /// Create a new environment for RocksDB.
+    pub fn new(path: impl AsRef<Path>) -> Result<RocksDbStore, rocksdb::Error> {
+        Self::new_with_opts(path, Self::rocksdb_defaults())
+    }
+
+    /// Create a new environment for RocksDB with the specified options.
+    pub fn new_with_opts(
+        path: impl AsRef<Path>,
+        rocksdb_opts: Options,
+    ) -> Result<RocksDbStore, rocksdb::Error> {
+        let db = Arc::new(DBWithThreadMode::<MultiThreaded>::open_cf(
+            &rocksdb_opts,
+            path.as_ref(),
+            ACTIVE_COLUMN_FAMILIES.to_vec(),
+        )?);
+
+        Ok(RocksDbStore {
+            db,
+            path: path.as_ref().to_path_buf(),
+        })
+    }
+
+    /// Default constructor for rocksdb options.
+    pub fn rocksdb_defaults() -> Options {
+        let mut factory_opts = BlockBasedOptions::default();
+        factory_opts.set_block_size(ROCKS_DB_BLOCK_SIZE_BYTES);
+
+        let mut db_opts = Options::default();
+        db_opts.set_block_based_table_factory(&factory_opts);
+
+        db_opts.set_compression_type(ROCKS_DB_COMPRESSION_TYPE);
+        db_opts.set_compression_options(
+            ROCKS_DB_WINDOW_BITS,
+            ROCKS_DB_ZSTD_COMPRESSION_LEVEL,
+            ROCKS_DB_ZSTD_STRATEGY,
+            ROCKS_DB_ZSTD_MAX_DICT_BYTES,
+        );
+
+        // seems to lead to a sporadic segfault within rocksdb compaction
+        // const ROCKS_DB_ZSTD_MAX_TRAIN_BYTES: i32 = 1024 * 1024; // 1 MB
+        // db_opts.set_zstd_max_train_bytes(ROCKS_DB_ZSTD_MAX_TRAIN_BYTES);
+
+        db_opts.set_compaction_style(ROCKS_DB_COMPACTION_STYLE);
+        db_opts.set_max_bytes_for_level_base(ROCKS_DB_MAX_LEVEL_FILE_SIZE_BYTES);
+        db_opts.set_max_open_files(ROCKS_DB_MAX_OPEN_FILES);
+
+        db_opts.create_missing_column_families(true);
+        db_opts.create_if_missing(true);
+
+        // recommended to match # of cores on host.
+        db_opts.increase_parallelism(num_cpus::get() as i32);
+
+        db_opts.create_missing_column_families(true);
+        db_opts.create_if_missing(true);
+
+        db_opts
+    }
+
+    /// Check if a state root has been marked as migrated from lmdb to rocksdb.
+    pub(crate) fn get_root_migration_state(
+        &self,
+        state_root: &[u8],
+    ) -> Result<RootMigration, error::Error> {
+        let lmdb_migration_column = self.lmdb_tries_migrated_column()?;
+        let migration_state = self.db.get_cf(&lmdb_migration_column, state_root)?;
+        Ok(match migration_state {
+            Some(state) if state.is_empty() => RootMigration::Complete,
+            Some(state) if state.get(0) == Some(&1u8) => RootMigration::Incomplete,
+            _ => RootMigration::NotMigrated,
+        })
+    }
+
+    /// Marks a state root as started migration from lmdb to rocksdb.
+    pub(crate) fn mark_state_root_migration_incomplete(
+        &self,
+        state_root: &[u8],
+    ) -> Result<(), error::Error> {
+        let lmdb_migration_column = self.lmdb_tries_migrated_column()?;
+        self.db.put_cf(&lmdb_migration_column, state_root, &[1u8])?;
+        Ok(())
+    }
+
+    /// Marks a state root as fully migrated from lmdb to rocksdb.
+    pub(crate) fn mark_state_root_migration_completed(
+        &self,
+        state_root: &[u8],
+    ) -> Result<(), error::Error> {
+        let lmdb_migration_column = self.lmdb_tries_migrated_column()?;
+        self.db.put_cf(&lmdb_migration_column, state_root, &[])?;
+        Ok(())
+    }
+
+    /// Trie V1 column family.
+    fn trie_column_family(&self) -> Result<Arc<BoundColumnFamily<'_>>, error::Error> {
+        self.db.cf_handle(TRIE_V1_COLUMN_FAMILY).ok_or_else(|| {
+            error::Error::UnableToOpenColumnFamily(TRIE_V1_COLUMN_FAMILY.to_string())
+        })
+    }
+
+    /// Column family tracking state roots migrated from lmdb (supports safely resuming if the node
+    /// were to be stopped during a migration).
+    fn lmdb_tries_migrated_column(&self) -> Result<Arc<BoundColumnFamily<'_>>, error::Error> {
+        self.db
+            .cf_handle(LMDB_MIGRATED_STATE_ROOTS_COLUMN_FAMILY)
+            .ok_or_else(|| {
+                error::Error::UnableToOpenColumnFamily(
+                    LMDB_MIGRATED_STATE_ROOTS_COLUMN_FAMILY.to_string(),
+                )
+            })
+    }
+
+    /// Working set column family.
+    pub fn working_set_column_family(&self) -> Result<Arc<BoundColumnFamily<'_>>, error::Error> {
+        self.db.cf_handle(WORKING_SET_COLUMN_FAMILY).ok_or_else(|| {
+            error::Error::UnableToOpenColumnFamily(WORKING_SET_COLUMN_FAMILY.to_string())
+        })
+    }
+
+    /// Trie store iterator.
+    pub fn trie_store_iterator<'a: 'b, 'b>(
+        &'a self,
+    ) -> Result<DBIteratorWithThreadMode<'b, DBWithThreadMode<MultiThreaded>>, error::Error> {
+        let cf_handle = self.trie_column_family()?;
+        Ok(self.db.iterator_cf(&cf_handle, IteratorMode::Start))
+    }
+
+    /// Return the path to the backing rocksdb files.
+    pub(crate) fn path(&self) -> PathBuf {
+        self.path.clone()
     }
 }
