+++ execution_engine/src/storage/trie_store/mod.rs
@@ -12,7 +12,5 @@ use casper_hashing::Digest;
 
 use crate::storage::{store::Store, trie::Trie};
 
-const NAME: &str = "TRIE_STORE";
-
 /// An entity which persists [`Trie`] values at their hashes.
 pub trait TrieStore<K, V>: Store<Digest, Trie<K, V>> {}
