+++ execution_engine/src/storage/trie_store/tests/simple.rs
@@ -6,9 +6,8 @@ use super::TestData;
 use crate::storage::{
     error::{self, in_memory},
     store::StoreExt,
-    transaction_source::{db::RocksDbStore, rocksdb_defaults},
     trie::Trie,
-    trie_store::{in_memory::InMemoryTrieStore, TrieStore},
+    trie_store::{db::RocksDbStore, in_memory::InMemoryTrieStore, TrieStore},
 };
 
 fn put_succeeds<K, V, S, E>(store: &S, items: &[TestData<K, V>]) -> Result<(), E>
@@ -33,7 +32,7 @@ fn in_memory_put_succeeds() {
 #[test]
 fn rocksdb_put_succeeds() {
     let tmp_dir = tempdir().unwrap();
-    let store = RocksDbStore::new(tmp_dir.path(), rocksdb_defaults()).unwrap();
+    let store = RocksDbStore::new(tmp_dir.path()).unwrap();
     let data = &super::create_data()[0..1];
     assert!(put_succeeds::<_, _, _, error::Error>(&store, data).is_ok());
     tmp_dir.close().unwrap();
@@ -76,7 +75,7 @@ fn in_memory_put_get_succeeds() {
 #[test]
 fn rocksdb_put_get_succeeds() {
     let tmp_dir = tempdir().unwrap();
-    let store = RocksDbStore::new(tmp_dir.path(), rocksdb_defaults()).unwrap();
+    let store = RocksDbStore::new(tmp_dir.path()).unwrap();
     let data = &super::create_data()[0..1];
 
     let expected: Vec<Trie<Bytes, Bytes>> = data.iter().cloned().map(|TestData(_, v)| v).collect();
@@ -113,7 +112,7 @@ fn in_memory_put_get_many_succeeds() {
 #[test]
 fn rocksdb_put_get_many_succeeds() {
     let tmp_dir = tempdir().unwrap();
-    let store = RocksDbStore::new(tmp_dir.path(), rocksdb_defaults()).unwrap();
+    let store = RocksDbStore::new(tmp_dir.path()).unwrap();
     let data = super::create_data();
 
     let expected: Vec<Trie<Bytes, Bytes>> = data.iter().cloned().map(|TestData(_, v)| v).collect();
