+++ resources/local/config.toml
@@ -372,13 +372,6 @@ get_from_peer_timeout = '3sec'
 # ========================================================
 [contract_runtime]
 
-# Optional maximum size of the database to use for the global state store.
-#
-# If unset, defaults to 32,212,254,720 == 30 GiB.
-#
-# The size should be a multiple of the OS page size.
-#max_global_state_size = 32_212_254_720
-
 # Optional depth limit to use for global state queries.
 #
 # If unset, defaults to 5.
