+++ utils/retrieve-state/src/storage.rs
@@ -1,19 +1,13 @@
 use std::{
-    env, fs,
+    env,
     path::{Path, PathBuf},
     sync::Arc,
 };
 
-use lmdb::DatabaseFlags;
-
 use casper_execution_engine::{
     core::engine_state::{EngineConfig, EngineState},
-    storage::{
-        global_state::db::DbGlobalState, transaction_source::db::LmdbEnvironment,
-        trie_store::db::LmdbTrieStore,
-    },
+    storage::{global_state::db::DbGlobalState, trie_store::db::RocksDbStore},
 };
-use casper_hashing::Digest;
 use casper_node::{
     storage::Storage,
     types::{Deploy, DeployHash},
@@ -21,9 +15,6 @@ use casper_node::{
 };
 use casper_types::{EraId, ProtocolVersion};
 use num_rational::Ratio;
-use tracing::info;
-
-use crate::DEFAULT_MAX_READERS;
 
 /// Gets many deploys by hash.
 pub fn get_many_deploys_by_hash(
@@ -45,29 +36,11 @@ pub fn get_many_deploys_by_hash(
     Ok(deploys)
 }
 
-/// Create an lmdb environment at a given path.
-fn create_lmdb_environment(
-    lmdb_path: impl AsRef<Path>,
-    default_max_db_size: usize,
-    manual_sync_enabled: bool,
-) -> Result<Arc<LmdbEnvironment>, anyhow::Error> {
-    let lmdb_environment = Arc::new(LmdbEnvironment::new(
-        &lmdb_path,
-        default_max_db_size,
-        DEFAULT_MAX_READERS,
-        manual_sync_enabled,
-    )?);
-    Ok(lmdb_environment)
-}
-
 /// Loads an existing execution engine.
 pub fn load_execution_engine(
     ee_lmdb_path: impl AsRef<Path>,
     rocksdb_path: impl AsRef<Path>,
-    default_max_db_size: usize,
-    state_root_hash: Digest,
-    manual_sync_enabled: bool,
-) -> Result<(Arc<EngineState<DbGlobalState>>, Arc<LmdbEnvironment>), anyhow::Error> {
+) -> Result<Arc<EngineState<DbGlobalState>>, anyhow::Error> {
     let lmdb_data_file = ee_lmdb_path.as_ref().join("data.lmdb");
     if !ee_lmdb_path.as_ref().join("data.lmdb").exists() {
         return Err(anyhow::anyhow!(
@@ -75,52 +48,25 @@ pub fn load_execution_engine(
             lmdb_data_file.display()
         ));
     }
-    let lmdb_environment =
-        create_lmdb_environment(&ee_lmdb_path, default_max_db_size, manual_sync_enabled)?;
-    let lmdb_trie_store = Arc::new(LmdbTrieStore::open(&lmdb_environment, None)?);
-    let global_state = DbGlobalState::new(
-        Arc::clone(&lmdb_environment),
-        lmdb_trie_store,
-        state_root_hash,
-        rocksdb_path,
+    let global_state = DbGlobalState::empty(
+        Some(ee_lmdb_path.as_ref().to_path_buf()),
+        RocksDbStore::new(rocksdb_path)?,
     )?;
-    Ok((
-        Arc::new(EngineState::new(global_state, EngineConfig::default())),
-        lmdb_environment,
-    ))
+    Ok(Arc::new(EngineState::new(
+        global_state,
+        EngineConfig::default(),
+    )))
 }
 
 /// Creates a new execution engine.
 pub fn create_execution_engine(
-    ee_lmdb_path: impl AsRef<Path>,
     rocksdb_path: impl AsRef<Path>,
-    default_max_db_size: usize,
-    manual_sync_enabled: bool,
-) -> Result<(Arc<EngineState<DbGlobalState>>, Arc<LmdbEnvironment>), anyhow::Error> {
-    if !ee_lmdb_path.as_ref().exists() {
-        info!(
-            "creating new lmdb data dir {}",
-            ee_lmdb_path.as_ref().display()
-        );
-        fs::create_dir_all(&ee_lmdb_path)?;
-    }
-    fs::create_dir_all(&ee_lmdb_path)?;
-    let lmdb_environment =
-        create_lmdb_environment(&ee_lmdb_path, default_max_db_size, manual_sync_enabled)?;
-    lmdb_environment.env().sync(true)?;
-
-    let lmdb_trie_store = Arc::new(LmdbTrieStore::new(
-        &lmdb_environment,
-        None,
-        DatabaseFlags::empty(),
-    )?);
-    let global_state =
-        DbGlobalState::empty(Arc::clone(&lmdb_environment), lmdb_trie_store, rocksdb_path)?;
-
-    Ok((
-        Arc::new(EngineState::new(global_state, EngineConfig::default())),
-        lmdb_environment,
-    ))
+) -> Result<Arc<EngineState<DbGlobalState>>, anyhow::Error> {
+    let global_state = DbGlobalState::empty(None, RocksDbStore::new(rocksdb_path.as_ref())?)?;
+    Ok(Arc::new(EngineState::new(
+        global_state,
+        EngineConfig::default(),
+    )))
 }
 
 pub fn normalize_path(path: impl AsRef<Path>) -> Result<PathBuf, anyhow::Error> {
